AVRO-1781 ADDENDUM: Remove LogicalTypes cache.
This cache maps a Schema to its corresponding LogicalType that has been
constructed from the Schema's properties, by object identity. This is no
longer needed now that the LogicalType is set on the schema instance
when it is parsed. The cache was necessary when we were trying to avoid
all changes to the Schema class.

git-svn-id: https://svn.apache.org/repos/asf/avro/trunk@1725668 13f79535-47bb-0310-9956-ffa450edef68